+++ TekSystems/Model/IReceipt.cs
@@ -2,10 +2,10 @@
 {
     public interface IReceipt
     {
-        double FinalPrice { get; set; }
-        double FinalPriceBeforeTaxes { get; set; }
+        decimal FinalPrice { get; set; }
+        decimal FinalPriceBeforeTaxes { get; set; }
         IShoppingCartItem[] SelectedItems { get; set; }
         int TotalItems { get; set; }
-        double TotalTaxes { get; set; }
+        decimal TotalTaxes { get; set; }
     }
 }
